Separate query from dataloader for pytorch
